+++ sfinder-lib/src/sfinder_core/neighbor/original_piece.rs
@@ -113,7 +113,7 @@ impl<'a> PartialEq for OriginalPiece<'a> {
 }
 
 impl HashCode for OriginalPiece<'_> {
-    type Output = u32;
+    type Output = u64;
 
     fn hash_code(&self) -> Self::Output {
         self.operation_with_key.hash_code()
